🚚(backend) move the token_generator module to services.token
The token generator is literally a service. It should stand at a proper
place.
